fixed "comparison between signed and unsigned" Signed-off-by: Eric Wong <normalperson@yhbt.net>
git-svn-id: https://svn.musicpd.org/mpd/trunk@7146 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/charConv.c
@@ -120,7 +120,7 @@ char *char_conv_str(char *dest, char *string)
 			    iconv(char_conv_iconv, &string, &inleft, &bufferPtr,
 				  &outleft);
 			if (outleft == BUFFER_SIZE
-			    || (err == -1L && errno != E2BIG)) {
+			    || (err == (size_t)-1L && errno != E2BIG)) {
 				return NULL;
 			}
 			memcpy(dest + retlen, buffer, BUFFER_SIZE - outleft);
